Moving the document list header component to components
